+++ CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [2.5.3] - 2018-11-17
+- Add travis job
+
+## [2.5.2] - 2018-11-02
+- Add the build to the repository
+
+## [2.5.1] - 2018-11-01
+### Improved
+- PASSBOLT-3188: As LU the UI shouldn't crash if a password uri cannot be parsed
+
+## [2.5.0] - 2018-10-30
 ### Added
 - PASSBOLT-3093: As LU I can select all passwords to perform a bulk operation
 - PASSBOLT-3121: MFA administration screen
@@ -75,8 +87,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - AP: User with plugin installed
 - LU: Logged in user
 
-[Unreleased]: https://github.com/passbolt/passbolt_api/compare/v2.4.0...HEAD
+[Unreleased]: https://github.com/passbolt/passbolt_api/compare/v2.5.3...HEAD
+[2.5.3]: https://github.com/passbolt/passbolt-appjs/compare/v2.5.2...v2.5.3
+[2.5.2]: https://github.com/passbolt/passbolt-appjs/compare/v2.5.1...v2.5.2
+[2.5.1]: https://github.com/passbolt/passbolt-appjs/compare/v2.5.0...v2.5.1
+[2.5.0]: https://github.com/passbolt/passbolt-appjs/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/passbolt/passbolt-appjs/compare/v2.3.0...v2.4.0
 [2.3.0]: https://github.com/passbolt/passbolt-appjs/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/passbolt/passbolt-appjs/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/passbolt/passbolt-appjs/compare/5df5207...v2.1.0
+
+
+"i-0d8164e6b9e73e34d" "i-08d005503ef1dcbc1" "i-0e5450f6a559cf4f8" "i-0eac703e3672af6a9" "i-0ae9b0523b36ca774" "i-0be1d2d995e7a23df" "i-0bb4922cfe2b7e8a3" "i-00ded96a8765a057d" "i-039dd790158235606" "i-020263d0befe41447"